requirements.txt
Flask
   requests
   python-dotenv
fuzzywuzzy[speedup]
geopy